pyproject.toml
[project]
name = "repl-nix-stockpriceforecast"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.3",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "scikit-learn>=1.5.2",
    "scipy>=1.14.1",
    "streamlit>=1.40.0",
    "xgboost>=2.1.2",
    "yfinance>=0.2.48",
]
